requirements.txt
# === ÉTAPE 1: DÉPENDANCES EXISTANTES (NE PAS TOUCHER) ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentinelhub==3.9.1
geopandas==0.14.1
numpy==1.24.3
matplotlib==3.7.2
python-multipart==0.0.6
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==21.2.0
Pillow==10.0.1
requests==2.31.0
fiona==1.9.5
shapely==2.0.2
pandas==2.1.3
pyproj==3.6.1
rasterio==1.3.9
scipy==1.11.4

# === ÉTAPE 2: NOUVELLES DÉPENDANCES (VERSIONS COMPATIBLES) ===
# OpenCV version headless pour Docker
opencv-python-headless==4.8.0.76

# scikit-image version compatible avec vos autres libs
scikit-image==0.20.0

# === ÉTAPE 3: SÉCURITÉ ET VALIDATION ===
# Pydantic pour validation stricte des données
pydantic==2.5.0

# Rate limiting pour FastAPI
slowapi==0.1.9

# Validation de fichiers
python-magic==0.4.27

# === ÉTAPE 4: OPTIMISATION DES PERFORMANCES ===
# Redis pour mise en cache
redis==5.0.1

# Traitement asynchrone amélioré
aiofiles==23.2.0

# Testing dependencies
httpx==0.28.1
